+++ crates/common/certificate/Cargo.toml
@@ -9,6 +9,8 @@ rust-version = "1.58.1"
 
 [dependencies]
 rcgen = { version = "0.9", features = ["pem", "zeroize"] }
+rustls_0_19 = {package = "rustls", version = "0.19.0" }
+rustls = "0.20.6"
 sha-1 = "0.10"
 thiserror = "1.0"
 time = "0.3"
@@ -20,4 +22,5 @@ anyhow = "1.0"
 assert_matches = "1.5"
 base64 = "0.13"
 pem = "1.0"
+tempfile = "3.2"
 time = {version = "0.3", features = ["macros"]}
