+++ crates/core/tedge/Cargo.toml
@@ -23,8 +23,6 @@ hyper = { version = "0.14", default-features = false }
 reqwest = { version = "0.11", default-features = false, features = ["blocking", "json", "rustls-tls", "stream"] }
 rpassword = "5.0"
 rumqttc = "0.10"
-rustls = "0.20.2"
-rustls_0_19 = {package = "rustls", version = "0.19.0" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tedge_config = { path = "../../common/tedge_config" }
